+++ src/mccode_antlr/common/expression/__init__.py
@@ -1,0 +1,12 @@
+from .types import ObjectType, ShapeType, DataType, OpStyle
+from .nodes import Op, TrinaryOp, BinaryOp, UnaryOp, Value, ExprNodeSingular, ExprNodeList, ExprNode, OpNode
+from .expr import Expr
+from .utils import unary_expr, binary_expr, value_or_op_from_dict, op_node_from_obj, op_post_init
+
+__all__ = [
+    'ObjectType', 'ShapeType', 'DataType', 'OpStyle',
+    'Op', 'TrinaryOp', 'BinaryOp', 'UnaryOp', 'Value',
+    'ExprNodeSingular', 'ExprNodeList', 'ExprNode', 'OpNode',
+    'Expr',
+    'unary_expr', 'binary_expr', 'value_or_op_from_dict', 'op_node_from_obj', 'op_post_init',
+]
