+++ src/mccode_antlr/common/expression/nodes.py
@@ -2,171 +2,10 @@ from __future__ import annotations
 
 from typing import Union, Optional
 
-from enum import Enum, IntEnum
-
 from loguru import logger
-
 from msgspec import Struct
 
-class ObjectType(Enum):
-    value = 1
-    initializer_list = 2
-    identifier = 3
-    function = 4
-    parameter = 5
-
-    def __str__(self):
-        return self.name
-
-    @staticmethod
-    def from_name(name):
-        if 'value' in name:
-            return ObjectType.value
-        if 'initializer_list' in name:
-            return ObjectType.initializer_list
-        if 'identifier' in name:
-            return ObjectType.identifier
-        if 'function' in name:
-            return ObjectType.function
-        if 'parameter' in name:
-            return ObjectType.parameter
-        raise RuntimeError(f"No known conversion from non-enumerated object type {name}")
-
-    @property
-    def is_id(self):
-        return self == ObjectType.identifier
-
-    @property
-    def is_parameter(self):
-        return self == ObjectType.parameter
-
-    @property
-    def is_function(self):
-        return self == ObjectType.function
-
-
-class ShapeType(Enum):
-    unknown = 0
-    scalar = 1
-    vector = 2
-
-    @property
-    def mccode_c_type(self):
-        return '*' if self.is_vector else ''
-
-    def compatible(self, other):
-        return self == ShapeType.unknown or other == ShapeType.unknown or self == other
-
-    @property
-    def is_scalar(self):
-        return self == ShapeType.scalar
-
-    @property
-    def is_vector(self):
-        return self == ShapeType.vector
-
-    def __str__(self):
-        return self.name
-
-    @staticmethod
-    def from_name(name):
-        if 'vector' in name:
-            return ShapeType.vector
-        if 'scalar' in name:
-            return ShapeType.scalar
-        return ShapeType.unknown
-
-
-class DataType(Enum):
-    undefined = 0
-    float = 1
-    int = 2
-    str = 3
-
-    def compatible(self, other):
-        if self == DataType.undefined or other == DataType.undefined or self == other:
-            return True
-        if (self == DataType.float and other == DataType.int) or (self == DataType.int and other == DataType.float):
-            return True
-        return False
-
-    # promotion rules:
-    def __add__(self, other):
-        if self == DataType.undefined:
-            return other
-        if other == DataType.undefined:
-            return self
-        if self == other:
-            return self
-        if (self == DataType.float and other == DataType.int) or (self == DataType.int and other == DataType.float):
-            return DataType.float
-        return DataType.str
-
-    __sub__ = __add__
-    __mul__ = __add__
-
-    def __truediv__(self, other):
-        if self == DataType.str or other == DataType.str:
-            raise RuntimeError('Division of strings is undefined')
-        return DataType.float
-
-    def __floordiv__(self, other):
-        return DataType.int
-
-    @classmethod
-    def from_name(cls, name):
-        if 'double' in name or 'float' in name:
-            return cls.float
-        if 'int' in name:
-            return cls.int
-        if 'char' in name or 'string' in name or 'str' in name:
-            return cls.str
-        return cls.undefined
-
-    @property
-    def name(self):
-        if self == DataType.int:
-            return 'int'
-        if self == DataType.float:
-            return 'float'
-        if self == DataType.str:
-            return 'str'
-        return 'undefined'
-
-    @property
-    def is_int(self):
-        return self == DataType.int
-
-    @property
-    def is_float(self):
-        return self == DataType.float
-
-    @property
-    def is_str(self):
-        return self == DataType.str
-
-    @property
-    def mccode_c_type(self):
-        if self == DataType.float:
-            return "double"
-        if self == DataType.int:
-            return "int"
-        if self == DataType.str:
-            return "char *"
-        raise RuntimeError(f"No known conversion from non-enumerated data type {self}")
-
-
-def _comb(f, s: list):
-    return ','.join(f(x) for x in s)
-
-
-def _fmt_comb(fmt, s: list):
-    return ','.join(f'{x:{fmt}}' for x in s)
-
-
-class OpStyle (IntEnum):
-    C = 1
-    PYTHON = 2
+from .types import DataType, ObjectType, OpStyle, ShapeType, _comb, _fmt_comb
 
 
 class Op(Struct, tag_field='type', tag='op'):
@@ -267,6 +106,7 @@ class Op(Struct, tag_field='type', tag='op'):
     def as_type(self, pdt):
         raise NotImplementedError()
 
+
 class TrinaryOp(Op, tag='trinary_op'):
     op: str
     first: OpNode
@@ -274,6 +114,7 @@ class TrinaryOp(Op, tag='trinary_op'):
     third: OpNode
 
     def __post_init__(self):
+        from .utils import op_post_init
         op_post_init(self, ['first', 'second', 'third'], {})
 
     def as_type(self, pdt):
@@ -356,6 +197,7 @@ class BinaryOp(Op, tag='binary_op'):
     right: OpNode
 
     def __post_init__(self):
+        from .utils import op_post_init
         special = {k: DataType.undefined for k in (
             '__call__', '__struct_access__', '__pointer_access__', '__getitem__')}
         op_post_init(self, ['left', 'right'], special)
@@ -443,9 +285,6 @@ class BinaryOp(Op, tag='binary_op'):
     @property
     def vector_known(self):
         return len(self.left) == 1 and len(self.right) == 1 and self.left[0].vector_known and self.right[0].vector_known
-    #
-    # def __len__(self):
-    #     return max(len(self.left), len(self.right))
 
     def simplify(self):
         left = [x.simplify() for x in self.left]
@@ -499,6 +338,7 @@ class UnaryOp(Op, tag='unary_op'):
     value: OpNode
 
     def __post_init__(self):
+        from .utils import op_post_init
         op_post_init(self, ['value'], {})
 
     def as_type(self, pdt):
@@ -556,9 +396,6 @@ class UnaryOp(Op, tag='unary_op'):
     def vector_known(self):
         return len(self.value) == 1 and self.value[0].vector_known
 
-    # def __len__(self):
-    #     return len(self.value)
-
     def __gt__(self, other):
         logger.debug(f'{self} > {other} has been called (but probably should not have been!)')
         return False
@@ -566,7 +403,7 @@ class UnaryOp(Op, tag='unary_op'):
     def simplify(self):
         value = [v.simplify() for v in self.value]
         if self.op == '__group__' and len(value) == 1 and isinstance(value[0], Value):
-            return value[0]  # Expr(value)
+            return value[0]
         elif self.op == '-' and len(value) == 1 and isinstance(value[0], Value):
             return -value[0]
         elif self.op == '+' and len(value) == 1 and isinstance(value[0], Value):
@@ -618,7 +455,6 @@ class Value(Struct, tag_field='type', tag='value'):
             if not isinstance(getattr(self, prop), typ):
                 setattr(self, prop, typ(getattr(self, prop)))
 
-
     def __int__(self):
         if self.data_type != DataType.int:
             raise RuntimeError('Non-integer data type Value; round first')
@@ -658,7 +494,6 @@ class Value(Struct, tag_field='type', tag='value'):
     @value.setter
     def value(self, value):
         logger.debug(f'Updating Value from {self._value} to {value}')
-        # self._object = self.data_type != DataType.str and isinstance(value, str)
         self._value = value
 
     @data_type.setter
@@ -699,6 +534,7 @@ class Value(Struct, tag_field='type', tag='value'):
         return hash(str(self))
 
     def compatible(self, other, id_ok=False):
+        from .expr import Expr
         if isinstance(other, Expr) and other.is_singular:
             other = other.expr[0]
         if isinstance(other, (UnaryOp, BinaryOp)):
@@ -731,11 +567,21 @@ class Value(Struct, tag_field='type', tag='value'):
         return cls(value, DataType.undefined, ObjectType.identifier, ShapeType.scalar)
 
     @classmethod
-    def array(cls, value, dt:DataType|None=None):
+    def parameter(cls, value, dt: DataType | None = None):
+        """Create a Value representing a known instrument parameter (ObjectType.parameter).
+
+        Unlike Value.id(), the ObjectType.parameter flag is set immediately so that
+        format(..., 'p') emits the _instrument_var._parameters. prefix without waiting
+        for verify_parameters() to be called.
+        """
+        return cls(value, dt if dt is not None else DataType.undefined, ObjectType.parameter, ShapeType.scalar)
+
+    @classmethod
+    def array(cls, value, dt: DataType | None = None):
         return cls(value, dt if dt is not None else DataType.undefined, None, ShapeType.vector)
 
     @classmethod
-    def function(cls, value, dt:DataType|None=None):
+    def function(cls, value, dt: DataType | None = None):
         return cls(value, dt if dt is not None else DataType.undefined, ObjectType.function)
 
     @classmethod
@@ -798,9 +644,6 @@ class Value(Struct, tag_field='type', tag='value'):
     def is_vector(self):
         return self.shape_type.is_vector
 
-    # def __len__(self):
-    #     return len(self.value) if self.is_vector else 1
-
     @property
     def vector_len(self):
         return len(self.value) if self.is_vector else 1
@@ -860,11 +703,6 @@ class Value(Struct, tag_field='type', tag='value'):
         pdt = self.data_type
         if other.is_op or self.is_id or other.is_id or pdt.is_str:
             return BinaryOp(self.data_type, OpStyle.C, '%', [self], [other])
-        # This is neither consistent with the Python or C definition
-        #   Python -- the result takes the sign of the divisor
-        #   C      -- the sign is always positive for positive numerator
-        #             but compiler-dependent for negative numerator
-        # return Value(self.value - (self.value // other.value) * other.value, pdt)
         return Value(self.value % other.value, pdt)
 
     def __truediv__(self, other):
@@ -915,6 +753,16 @@ class Value(Struct, tag_field='type', tag='value'):
         if other.is_op:
             return False
         return self.value == other.value
+
+    def eq(self, other) -> 'BinaryOp':
+        """Return a BinaryOp expression node for `self == other` (never a bool)."""
+        other = other if isinstance(other, (Value, Op)) else Value.best(other)
+        return BinaryOp(self.data_type, OpStyle.C, '__eq__', [self], [other])
+
+    def ne(self, other) -> 'BinaryOp':
+        """Return a BinaryOp expression node for `self != other`."""
+        other = other if isinstance(other, (Value, Op)) else Value.best(other)
+        return BinaryOp(self.data_type, OpStyle.C, '__neq__', [self], [other])
 
     def __lt__(self, other):
         other = other if isinstance(other, (Value, Op)) else Value.best(other)
@@ -974,6 +822,7 @@ class Value(Struct, tag_field='type', tag='value'):
 
     def evaluate(self, known: dict):
         if not self.is_constant and self.value in known:
+            from .expr import Expr
             result = known[self.value]
             if isinstance(result, Expr) and result.is_singular:
                 return result.expr[0]
@@ -1001,443 +850,8 @@ class Value(Struct, tag_field='type', tag='value'):
             self._object = ObjectType.parameter
 
 
+# Type aliases used by Op subclasses and Expr
 ExprNodeSingular = Union[Value, UnaryOp, BinaryOp, TrinaryOp]
 ExprNodeList = list[ExprNodeSingular]
 ExprNode = Union[ExprNodeSingular, ExprNodeList]
 OpNode = list[Union[Value, UnaryOp, BinaryOp, TrinaryOp, Op]]
-
-
-class Expr(Struct):
-    expr: ExprNode
-
-    @classmethod
-    def from_dict(cls, args: dict):
-        expr = args['expr']
-        if not hasattr(expr, '__len__'):
-            expr = [expr]
-        return cls([value_or_op_from_dict(x) for x in expr])
-
-    def __post_init__(self):
-        if not isinstance(self.expr, list):
-            self.expr = [self.expr]
-        if any(not isinstance(node, ExprNodeSingular) for node in self.expr):
-            types = list(dict.fromkeys(type(node) for node in self.expr).keys())
-            raise ValueError(f"An Expr can not be a list of {types}")
-
-    def __str__(self):
-        return ','.join(str(x) for x in self.expr)
-
-    def __format__(self, format_spec):
-        """Abuse string formatting to append the _instrument_var.parameters prefix to parameter names"""
-        return ','.join(format(x, format_spec) for x in self.expr)
-
-    def __repr__(self):
-        return ','.join(repr(x) for x in self.expr)
-
-    def __hash__(self):
-        return hash(str(self))
-
-    def __contains__(self, value):
-        return any(value in x for x in self.expr)
-
-    @staticmethod
-    def parse(s: str):
-        from antlr4 import InputStream
-        from ..grammar import McInstr_parse
-        from ..instr import InstrVisitor
-        visitor = InstrVisitor(None, None)
-        return visitor.getExpr(McInstr_parse(InputStream(s), 'expr'))
-
-    @classmethod
-    def float(cls, value):
-        if isinstance(value, cls):
-            # Make sure the held expression *thinks* it's a float:
-            for expr in value.expr:
-                if expr.data_type != DataType.float:
-                    expr.data_type = DataType.float
-            return value
-        return cls(Value.float(value))
-
-    @classmethod
-    def int(cls, value):
-        if isinstance(value, cls):
-            # Make sure the held expression *thinks* it's a float:
-            for expr in value.expr:
-                if expr.data_type != DataType.int:
-                    expr.data_type = DataType.int
-            return value
-        return cls(Value.int(value))
-
-    @classmethod
-    def str(cls, value):
-        if isinstance(value, cls):
-            # Make sure the held expression *thinks* it's a float:
-            for expr in value.expr:
-                if expr.data_type != DataType.str:
-                    expr.data_type = DataType.str
-            return value
-        return cls(Value.str(value))
-
-    @classmethod
-    def id(cls, value):
-        if isinstance(value, cls):
-            return value
-        return cls(Value.id(value))
-
-    @classmethod
-    def array(cls, value):
-        return cls(Value.array(value))
-
-    @classmethod
-    def function(cls, value):
-        return cls(Value.function(value))
-
-    @classmethod
-    def best(cls, value):
-        return cls(Value.best(value))
-
-    @property
-    def is_singular(self):
-        return len(self.expr) == 1
-
-    @property
-    def is_op(self):
-        return len(self.expr) == 1 and self.expr[0].is_op
-
-    @property
-    def is_zero(self):
-        return len(self.expr) == 1 and self.expr[0].is_zero
-
-    @property
-    def is_id(self):
-        return len(self.expr) == 1 and self.expr[0].is_id
-
-    @property
-    def is_parameter(self):
-        return len(self.expr) == 1 and self.expr[0].is_parameter
-
-    @property
-    def is_str(self):
-        return len(self.expr) == 1 and self.expr[0].is_str
-
-    @property
-    def is_scalar(self):
-        return len(self.expr) == 1 and self.expr[0].is_scalar
-
-    def is_value(self, value):
-        return len(self.expr) == 1 and self.expr[0].is_value(value)
-
-    @property
-    def is_vector(self):
-        return len(self.expr) == 1 and self.expr[0].is_vector
-    #
-    # def __len__(self):
-    #     return len(self.expr)
-    @property
-    def vector_len(self):
-        if len(self.expr) != 1:
-            raise RuntimeError('No vector_len for array Expr objects')
-        return self.expr[0].vector_len
-
-    @property
-    def is_constant(self):
-        return len(self.expr) == 1 and isinstance(self.expr[0], Value) and not self.expr[0].is_id
-
-    @property
-    def has_value(self):
-        return self.is_constant and self.expr[0].has_value
-
-    @property
-    def vector_known(self):
-        return self.is_constant and self.expr[0].vector_known
-
-    @property
-    def value(self):
-        if not self.is_constant:
-            raise NotImplementedError("No conversion from expressions to constants ... yet")
-        return self.expr[0].value
-
-    @property
-    def first(self):
-        return self.expr[0]
-
-    @property
-    def last(self):
-        return self.expr[-1]
-
-    def compatible(self, other, id_ok=False):
-        other_expr = other.expr[0] if (isinstance(other, Expr) and len(other.expr) == 1) else other
-        return len(self.expr) == 1 and self.expr[0].compatible(other_expr, id_ok)
-
-    def _prep_numeric_operation(self, msg: str, other):
-        if len(self.expr) != 1:
-            raise RuntimeError(f'Can not {msg} array Expr')
-        return other.expr[0] if (isinstance(other, Expr) and len(other.expr) == 1) else other
-
-    def _prep_rev_numeric_operation(self, msg: str, other):
-        r = self._prep_numeric_operation(msg, other)
-        if not isinstance(r, (Expr, TrinaryOp, BinaryOp, UnaryOp, Value)):
-            r = Value.best(r)
-        return r
-
-    def __add__(self, other):
-        return Expr(self.expr[0] + self._prep_numeric_operation('add to', other))
-
-    def __sub__(self, other):
-        return Expr(self.expr[0] - self._prep_numeric_operation('subtract', other))
-
-    def __mul__(self, other):
-        return Expr(self.expr[0] * self._prep_numeric_operation('multiply', other))
-
-    def __mod__(self, other):
-        return Expr(self.expr[0] % self._prep_numeric_operation('mod', other))
-
-    def __truediv__(self, other):
-        return Expr(self.expr[0] / self._prep_numeric_operation('divide', other))
-
-    def __floordiv__(self, other):
-        return Expr(self.expr[0] // self._prep_numeric_operation('divide', other))
-
-    def __pow__(self, other):
-        return Expr(self.expr[0] ** self._prep_numeric_operation('raise', other))
-
-    def __radd__(self, other):
-        return Expr(self._prep_rev_numeric_operation('add to', other) + self.expr[0])
-
-    def __rsub__(self, other):
-        return Expr(self._prep_rev_numeric_operation('subtract', other) - self.expr[0])
-
-    def __rmul__(self, other):
-        return Expr(self._prep_rev_numeric_operation('multiply', other) * self.expr[0])
-
-    def __rtruediv__(self, other):
-        return Expr(self._prep_rev_numeric_operation('divide', other) / self.expr[0])
-
-    def __rfloordiv__(self, other):
-        return Expr(self._prep_rev_numeric_operation('divide', other) // self.expr[0])
-
-    def __rpow__(self, other):
-        return Expr(self._prep_rev_numeric_operation('raise', other) ** self.expr[0])
-
-    def __neg__(self):
-        return Expr([-x for x in self.expr])
-
-    def __pos__(self):
-        return self
-
-    def __abs__(self):
-        return Expr([abs(x) for x in self.expr])
-
-    def __round__(self, n=None):
-        return Expr([round(x, n) for x in self.expr])
-
-    def __eq__(self, other):
-        if isinstance(other, Expr):
-            if len(other.expr) != len(self.expr):
-                return False
-            for o_expr, s_expr in zip(other.expr, self.expr):
-                if o_expr != s_expr:
-                    return False
-            return True
-        return len(self.expr) == 1 and self.expr[0] == other
-
-    def __lt__(self, other):
-        if isinstance(other, Expr):
-            if len(other.expr) != len(self.expr):
-                raise RuntimeError('Can not compare unequal-sized-array Expr objects')
-            for o_expr, s_expr in zip(other.expr, self.expr):
-                if o_expr <= s_expr:
-                    return False
-            return True
-        return len(self.expr) == 1 and self.expr[0] < other
-
-    def __gt__(self, other):
-        if isinstance(other, Expr):
-            if len(other.expr) != len(self.expr):
-                raise RuntimeError('Can not compare unequal-sized-array Expr objects')
-            for o_expr, s_expr in zip(other.expr, self.expr):
-                if o_expr >= s_expr:
-                    return False
-            return True
-        return len(self.expr) == 1 and self.expr[0] > other
-
-    def __le__(self, other):
-        if isinstance(other, Expr):
-            if len(other.expr) != len(self.expr):
-                raise RuntimeError('Can not compare unequal-sized-array Expr objects')
-            for o_expr, s_expr in zip(other.expr, self.expr):
-                if o_expr < s_expr:
-                    return False
-            return True
-        return len(self.expr) == 1 and self.expr[0] <= other
-
-    def __ge__(self, other):
-        if isinstance(other, Expr):
-            if len(other.expr) != len(self.expr):
-                raise RuntimeError('Can not compare unequal-sized-array Expr objects')
-            for o_expr, s_expr in zip(other.expr, self.expr):
-                if o_expr > s_expr:
-                    return False
-            return True
-        return len(self.expr) == 1 and self.expr[0] >= other
-
-    def __int__(self):
-        if not len(self.expr) == 1:
-            raise RuntimeError('No conversion to int for array Expr objects')
-        return int(self.expr[0])
-
-    @property
-    def mccode_c_type(self):
-        if len(self.expr) != 1:
-            raise RuntimeError('No McCode C type for array Expr objects')
-        if not isinstance(self.expr[0], Value):
-            logger.critical(f'Why is {self.expr[0]} not a Value?')
-        return self.expr[0].mccode_c_type
-
-    @property
-    def mccode_c_type_name(self):
-        if len(self.expr) != 1:
-            raise RuntimeError('No McCode C type name for array Expr objects')
-        return self.expr[0].mccode_c_type_name
-
-    @property
-    def data_type(self):
-        if len(self.expr) != 1:
-            raise RuntimeError('No data type for array Expr objects')
-        return self.expr[0].data_type
-
-    @data_type.setter
-    def data_type(self, dt):
-        if len(self.expr) != 1:
-            raise RuntimeError('No data type for array Expr objects')
-        self.expr[0].data_type = dt
-
-    @property
-    def shape_type(self):
-        if len(self.expr) != 1:
-            raise RuntimeError('No data type for array Expr objects')
-        return self.expr[0].shape_type
-
-    @shape_type.setter
-    def shape_type(self, st):
-        if len(self.expr) != 1:
-            raise RuntimeError('No data type for array Expr objects')
-        if not isinstance(self.expr[0], Value):
-            raise RuntimeError('No data type for non-scalar-Value Expr objects')
-        self.expr[0].shape_type = st
-
-    def simplify(self):
-        """Perform a very basic analysis to reduce the expression complexity"""
-        def simplify_to_single_or_list(node):
-            s = node.simplify()
-            return s[0] if hasattr(s, '__len__') and len(s) == 1 else s
-        return Expr([simplify_to_single_or_list(x) for x in self.expr])
-
-    def evaluate(self, known: dict):
-        def evaluate_to_single_or_list(node):
-            s = node.evaluate(known)
-            return s[0] if hasattr(s, '__len__') and len(s) == 1 else s
-        return Expr([evaluate_to_single_or_list(x) for x in self.expr]).simplify()
-
-    def depends_on(self, name: str):
-        return any(x.depends_on(name) for x in self.expr)
-
-    def copy(self):
-        return Expr([x.copy() for x in self.expr])
-
-    def verify_parameters(self, instrument_parameter_names: list[str]):
-        for x in self.expr:
-            x.verify_parameters(instrument_parameter_names)
-
-
-def unary_expr(func, name, v):
-    ops = {'cos': 'acos', 'sin': 'asin', 'tan': 'atan'}
-    if isinstance(v, Expr):
-        v = v.expr
-    if len(v) == 1:
-        v = v[0]
-    if isinstance(v, UnaryOp) and ((name in ops and v.op == ops[name]) or (v.op in ops and name == ops[v.op])):
-        return Expr(v.value)
-    if isinstance(v, Value) and not v.is_id:
-        if v.is_str or isinstance(v.value, str):
-            raise RuntimeError(f'How is a _string_ valued parameter, {v} not an identifier?')
-        return Expr(Value.best(func(v.value)))
-    return Expr(UnaryOp(DataType.float, OpStyle.C, name, v))
-
-
-def binary_expr(func, name, left, right):
-    ops = {'atan2': ('sin', 'cos')}
-    if isinstance(left, Expr):
-        left = left.expr
-    if isinstance(right, Expr):
-        right = right.expr
-    if len(left) == 1:
-        left = left[0]
-    if len(right) == 1:
-        right = right[0]
-    if isinstance(left, UnaryOp) and isinstance(right, UnaryOp) and name in ops:
-        left_func, right_func = ops[name]
-        if left.op == left_func and right.op == right_func and left.value == right.value:
-            return Expr(left.value)
-    if isinstance(left, Value) and isinstance(right, Value) and not left.is_id and not right.is_id:
-        return Expr(Value.best(func(left.value, right.value)))
-    return Expr(BinaryOp(DataType.float, OpStyle.C, name, left, right))
-
-
-
-def value_or_op_from_dict(args: dict):
-    # Strip the msgspec tag discriminator if present (added by tag_field='type')
-    args = {k: v for k, v in args.items() if k != 'type'}
-    # Value: _value, _data, _object, _shape,
-    if all(x in args for x in ('_value', '_data', '_object', '_shape')):
-        return Value(**args)
-    # UnaryOp, BinaryOp, TrinaryOp
-    if all(x in args for x in ('data_type', 'style', 'op')):
-        if all(x in args for x in ('value',)):
-            return UnaryOp(**args)
-        if all(x in args for x in ('left', 'right')):
-            return BinaryOp(**args)
-        if all(x in args for x in ('first', 'second', 'third')):
-            return TrinaryOp(**args)
-    print(f'{args=}')
-    raise ValueError("Not a valid expression node element")
-
-
-def op_node_from_obj(obj):
-    if not hasattr(obj, '__len__'):
-        raise ValueError('Op nodes should have length')
-    return [value_or_op_from_dict(d) for d in obj]
-
-
-def op_post_init(obj, props: list[str], special: dict):
-    for prop in props:
-        if isinstance(getattr(obj, prop), Expr):
-            setattr(obj, prop, getattr(obj, prop).expr)
-        if not isinstance(getattr(obj, prop), list):
-            setattr(obj, prop, [getattr(obj, prop)])
-
-    def data_type_set_list(i):
-        return list(dict.fromkeys(x.data_type for x in getattr(obj, props[i])))
-
-    for prop in props:
-        if any(isinstance(x, dict) for x in getattr(obj, prop)):
-            setattr(obj, prop, op_node_from_obj(getattr(obj, prop)))
-
-    if obj.op in special:
-        obj.data_type = special[obj.op]
-    else:
-        # In C the data types for `if_true` and `if_false` in
-        #   `(logical test) ? if_true : if_false;`
-        # _must_ be the same. Python is more flexible, but we need to at least have a promotable common type
-        # for compatibility with C. (the type of the logical test can differ)
-        a = data_type_set_list(-1)
-        b = data_type_set_list(-2) if len(props) > 1 else a
-        if len(a) != 1 or len(b) != 1:
-            raise RuntimeError('Multiple data types in one value not supported')
-        # datatype + datatype promotion preserves A+A=A
-        obj.data_type = a[0] + b[0]
-
-    # ensure that data_type is a DataType and style is an OpStyle?
-    for prop, typ in (('data_type', DataType), ('style', OpStyle)):
-        if not isinstance(p:=getattr(obj, prop), typ):
-            setattr(obj, prop, typ(p))
